backend/requirements.txt
fastapi[standard]>=0.115,<1.0
uvicorn>=0.34,<1.0
PyJWT[crypto]>=2.10,<3.0
httpx>=0.28,<1.0
pydantic-settings>=2.9,<3.0
